+++ src/roq/kucoin_futures/drop_copy.h
@@ -61,6 +61,7 @@ class DropCopy final : public core::web::Socket::Handler, public json::Parser::H
   void operator()(const core::web::Socket::Close &) override;
   void operator()(const core::web::Socket::Latency &) override;
   void operator()(const core::web::Socket::Text &) override;
+  void operator()(const core::web::Socket::Binary &) override;
 
  private:
   void operator()(ConnectionStatus);