+++ src/roq/kucoin_futures/market_data.h
@@ -71,6 +71,7 @@ class MarketData final : public core::web::Socket::Handler, public json::Parser:
   void operator()(const core::web::Socket::Close &) override;
   void operator()(const core::web::Socket::Latency &) override;
   void operator()(const core::web::Socket::Text &) override;
+  void operator()(const core::web::Socket::Binary &) override;
 
  private:
   void operator()(ConnectionStatus);